数据库连接跑通，初步单元测试跑通

+++ src/models/dream.py
@@ -1,5 +1,5 @@
 from libs.database.mysql import BaseModel
-from sqlalchemy import Column, String, Text, INTEGER, DateTime
+from sqlalchemy import Column, String, Text, Integer, DateTime
 
 
 class Dream(BaseModel):
@@ -9,6 +9,6 @@ class Dream(BaseModel):
     user_id = Column(String(40), nullable=False)
     title = Column(String(400), nullable=False)
     word = Column(Text, nullable=False)
-    status = Column(INTEGER(11), nullable=False)
+    status = Column(Integer, nullable=False)
     remark = Column(String(400))
     created = Column(DateTime, nullable=False)
